Move push_registrations to notify platform file
