+++ homeassistant/components/mobile_app/__init__.py
@@ -5,11 +5,10 @@ from homeassistant.components.webhook import async_register as webhook_register
 from homeassistant.helpers import device_registry as dr, discovery
 from homeassistant.helpers.typing import ConfigType, HomeAssistantType
 
-from .const import (ATTR_APP_DATA, ATTR_DEVICE_ID, ATTR_DEVICE_NAME,
-                    ATTR_MANUFACTURER, ATTR_MODEL, ATTR_PUSH_TOKEN,
-                    ATTR_PUSH_URL, ATTR_OS_VERSION, DATA_BINARY_SENSOR,
-                    DATA_CONFIG_ENTRIES, DATA_DELETED_IDS, DATA_DEVICES,
-                    DATA_SENSOR, DATA_STORE, DOMAIN, STORAGE_KEY,
+from .const import (ATTR_DEVICE_ID, ATTR_DEVICE_NAME,
+                    ATTR_MANUFACTURER, ATTR_MODEL, ATTR_OS_VERSION,
+                    DATA_BINARY_SENSOR, DATA_CONFIG_ENTRIES, DATA_DELETED_IDS,
+                    DATA_DEVICES, DATA_SENSOR, DATA_STORE, DOMAIN, STORAGE_KEY,
                     STORAGE_VERSION)
 
 from .http_api import RegistrationsView
@@ -118,15 +117,3 @@ class MobileAppFlowHandler(config_entries.ConfigFlow):
         """Handle a flow initialized during registration."""
         return self.async_create_entry(title=user_input[ATTR_DEVICE_NAME],
                                        data=user_input)
-
-
-def push_registrations(hass):
-    """Return a dictionary of push enabled registrations."""
-    targets = {}
-    for webhook_id, entry in hass.data[DOMAIN][DATA_CONFIG_ENTRIES].items():
-        data = entry.data
-        app_data = data[ATTR_APP_DATA]
-        if ATTR_PUSH_TOKEN in app_data and ATTR_PUSH_URL in app_data:
-            device_name = data[ATTR_DEVICE_NAME]
-            targets[device_name] = webhook_id
-    return targets
